// src/Constants/Enum.ts
export enum nameCity {
  NewYorkCity = 0,
  CapeTownSouthAfrica = 1,
  BerlinGermany = 2,
  LondonUK = 3,
  MadridSpain = 4,
  MexicoCityMexico = 5,
  LiverpoolUK = 6,
  TokyoJapan = 7,
  RomeItaly = 8,
  PortoPortugal = 9,
  ParisFrance = 10,
  MumbaiIndia = 11,
  LisbonPortugal = 12,
  Chicago = 13,
  ManchesterUK = 14,
  SaoPauloBrazil = 15,
  LosAngeles = 16,
  AmsterdamTheNetherlands = 17,
  Lagos = 18,
  MelbourneAustralia = 19,
}

export enum isStatus {
  statusPush = 1,
  statusProblem = 2,
  statusReady = 3,
}

export enum valueMoney {
  namtram = 0,
  motnghin = 1,
  motnghinnamtram = 1,
  hainghin = 1,
  honhainghin = 1,
}